refactor(orchestrator): extract TextStreamingMixin from the God-file (D-164 Slice 2)
Move the markdown typing-effect streamer (_stream_markdown_typing, LaTeX-preserving),
_format_history_for_prompt, _strip_retrieval_tags / _sanitize_text_for_user (+ the
_RETRIEVAL_TAG_RE it owns) verbatim into a cohesive TextStreamingMixin (SRP). All call sites
resolve through the MRO — byte-identical behaviour. Append to TUTOR_SOURCE_FILES (one line);
no source-inspection test needed conversion (all references are call sites in chat_with_agent
that stay put). God-file 3,678 -> 3,521 lines.
